a litle bit design change, could not finish it because android studio 3.4.1 has bags. I need to reinstall it

+++ app/src/main/java/com/example/sonyadmin/bar/product/Product.kt
@@ -1,6 +1,6 @@
 package com.example.sonyadmin.bar.product
 
 import com.example.sonyadmin.bar.category.Category
-data class ProductDetails(var amount:Long , var price: Long ,var volume : Double = 1.0)
+data class ProductDetails(var amount:Long , var price: Double ,var volume : Double = 1.0)
 data class Product(var name : String , var details : ProductDetails)
 data class Products(var list:List<Product>)
